requirements.txt
streamlit
python-jose[cryptography]